requirements.txt
streamlit==1.32.0
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4
Flask==3.0.2
Werkzeug==3.0.1
requests
tensorflow
onnxruntime
tqdm
h5py